fix click out behavior; style dialog;
previously dialog would close when, after manipulating an input (e.g.
range slider), if you lifted the pointer outside of the dialog (as in
the case where you drag past the end of the range input) the dialog
would close.

this was disconcerting, so I changed it to only close on drag out if the
target of the action is the dialog itself, and not one of it's children.