don't allow dragging cursor by toggle button
enabling the use to drag the cursor around by the show/hide controls
toggle means that they can bump the cursor out of place when toggling
the controls, which is undesirable. Also, it pollutes the undo stack
with lots of minor setCursor events (nudging by negligible values) which
makes the undoing experience worse.